+++ test/test_source_multi.py
@@ -74,7 +74,6 @@ def testTilesFromMultiSimpleScaling():
 
 
 @pytest.mark.skipif(sys.version_info < (3, 7), reason='requires python >= 3.7 for a sub-source')
-@pytest.mark.skipif(sys.version_info >= (3, 11), reason='requires python3.11 wheels')
 def testTilesFromMultiMultiSource(multiSourceImagePath):
     imagePath = multiSourceImagePath
     source = large_image_source_multi.open(imagePath)
@@ -125,7 +124,6 @@ def testTilesFromMultiString():
 
 
 @pytest.mark.skipif(sys.version_info < (3, 7), reason='requires python >= 3.7 for a sub-source')
-@pytest.mark.skipif(sys.version_info >= (3, 11), reason='requires python3.11 wheels')
 def testInternalMetadata(multiSourceImagePath):
     imagePath = multiSourceImagePath
     source = large_image_source_multi.open(imagePath)
@@ -134,7 +132,6 @@ def testInternalMetadata(multiSourceImagePath):
 
 
 @pytest.mark.skipif(sys.version_info < (3, 7), reason='requires python >= 3.7 for a sub-source')
-@pytest.mark.skipif(sys.version_info >= (3, 11), reason='requires python3.11 wheels')
 def testAssociatedImages(multiSourceImagePath):
     imagePath = multiSourceImagePath
     source = large_image_source_multi.open(imagePath)
@@ -194,7 +191,6 @@ def testFramesAsAxes():
 
 
 @pytest.mark.skipif(sys.version_info < (3, 7), reason='requires python >= 3.7 for a sub-source')
-@pytest.mark.skipif(sys.version_info >= (3, 11), reason='requires python3.11 wheels')
 def testMultiComposite():
     datastore.fetch('ITGA3Hi_export_crop2.nd2')
     imagePath = datastore.fetch('multi-source-composite.yaml')